+++ skbio/stats/tests/test_gradient.py
@@ -13,7 +13,7 @@ from unittest import TestCase, main
 import numpy as np
 import pandas as pd
 import numpy.testing as npt
-import pandas.util.testing as pdt
+import pandas.testing as pdt
 
 from skbio.util import get_data_path, assert_data_frame_almost_equal
 from skbio.stats.gradient import (GradientANOVA, AverageGradientANOVA,
@@ -254,14 +254,14 @@ class GradientTests(BaseTests):
         w_vector = pd.Series(np.array([1, 2, 3, 4, 5, 6, 7, 8]),
                              ['s1', 's2', 's3', 's4',
                               's5', 's6', 's7', 's8']).astype(np.float64)
-        exp = pd.DataFrame.from_dict({'s1': np.array([1.0]),
-                                      's2': np.array([2.0]),
-                                      's3': np.array([3.0]),
-                                      's4': np.array([4.0]),
-                                      's5': np.array([5.0]),
-                                      's6': np.array([6.0]),
-                                      's7': np.array([7.0]),
-                                      's8': np.array([8.0])
+        exp = pd.DataFrame.from_dict({'s1': np.array([1]),
+                                      's2': np.array([2]),
+                                      's3': np.array([3]),
+                                      's4': np.array([4]),
+                                      's5': np.array([5]),
+                                      's6': np.array([6]),
+                                      's7': np.array([7]),
+                                      's8': np.array([8])
                                       },
                                      orient='index')
         obs = _weight_by_vector(trajectory, w_vector)
@@ -276,11 +276,11 @@ class GradientTests(BaseTests):
         trajectory.sort_values(by=0, inplace=True)
         w_vector = pd.Series(np.array([25, 30, 35, 40, 45]),
                              ['s2', 's3', 's4', 's5', 's6']).astype(np.float64)
-        exp = pd.DataFrame.from_dict({'s2': np.array([2.0]),
-                                      's3': np.array([3.0]),
-                                      's4': np.array([4.0]),
-                                      's5': np.array([5.0]),
-                                      's6': np.array([6.0])}, orient='index')
+        exp = pd.DataFrame.from_dict({'s2': np.array([2]),
+                                      's3': np.array([3]),
+                                      's4': np.array([4]),
+                                      's5': np.array([5]),
+                                      's6': np.array([6])}, orient='index')
         obs = _weight_by_vector(trajectory, w_vector)
         assert_data_frame_almost_equal(obs.sort_index(), exp.sort_index())
 
@@ -296,7 +296,7 @@ class GradientTests(BaseTests):
                                       's2': np.array([2, 3, 4]),
                                       's3': np.array([5, 6, 7]),
                                       's4': np.array([8, 9, 10])},
-                                     orient='index').astype(np.float64)
+                                     orient='index')
         obs = _weight_by_vector(trajectory, w_vector)
         assert_data_frame_almost_equal(obs.sort_index(), exp.sort_index())
